+++ nextflow.config
@@ -132,20 +132,13 @@ profiles {
         params.run_docker = false
         params.run_singularity = false
 
-        conda.useMamba = true
-
-        process {
-            withLabel: main {
-                conda.enabled = true
-                docker.enabled = false
-                singularity.enabled = false
-            }
-            withLabel: vadr {
-                conda.enabled = true
-                docker.enabled = true
-                singularity.enabled = false
-            }
+        conda.enabled = true
+        if ( params.run_vadr == true ) {
+            docker.enabled = true
         }
+        singularity.enabled = false
+
+        conda.useMamba = true
     }
 
     // run singularity  + specify settings 
@@ -153,6 +146,12 @@ profiles {
         params.run_singularity = true
         params.run_docker = false 
         params.run_conda = false
+
+        conda.enabled = false
+        if ( params.run_vadr == true ) {
+            docker.enabled = true
+        }
+        singularity.enabled = true
 
         singularity.autoMounts = false
         singularity.cacheDir = "$PWD"
@@ -167,15 +166,7 @@ profiles {
 
         process {
             withLabel: main {
-                conda.enabled = false
-                docker.enabled = false
-                singularity.enabled = true
                 container = "$projectDir/app/singularity/containers/singularity_docker.sif"
-            }
-            withLabel: vadr {
-                conda.enabled = false
-                docker.enabled = true
-                singularity.enabled = true
             }
         }
     }
